refactor: move executor tests to executor.rs
