+++ dapp/src/executor.rs
@@ -91,3 +91,166 @@ pub fn initialize_contracts<T: IntoIterator<Item = (Address, Bytes)>>(contracts:
         })
         .collect::<BTreeMap<_, _>>()
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::{
+        decode_revert,
+        test_helpers::{new_backend, new_vicinity, COMPILED},
+    };
+    use dapp_utils::get_func;
+
+    use ethers::utils::id;
+    use evm::{ExitReason, ExitRevert, ExitSucceed};
+
+    #[test]
+    fn can_call_vm_directly() {
+        let cfg = Config::istanbul();
+        let compiled = COMPILED.get("Greeter").expect("could not find contract");
+
+        let addr = "0x1000000000000000000000000000000000000000"
+            .parse()
+            .unwrap();
+        let state = initialize_contracts(vec![(addr, compiled.runtime_bytecode.clone())]);
+
+        let vicinity = new_vicinity();
+        let backend = new_backend(&vicinity, state);
+        let mut dapp = Executor::new(12_000_000, &cfg, &backend);
+
+        let (_, status, _) = dapp
+            .call::<(), _>(
+                Address::zero(),
+                addr,
+                &get_func("function greet(string greeting) external").unwrap(),
+                "hi".to_owned(),
+                0.into(),
+            )
+            .unwrap();
+        assert_eq!(status, ExitReason::Succeed(ExitSucceed::Stopped));
+
+        let (retdata, status, _) = dapp
+            .call::<String, _>(
+                Address::zero(),
+                addr,
+                &get_func("function greeting() public view returns (string)").unwrap(),
+                (),
+                0.into(),
+            )
+            .unwrap();
+        assert_eq!(status, ExitReason::Succeed(ExitSucceed::Returned));
+        assert_eq!(retdata, "hi");
+    }
+
+    #[test]
+    fn solidity_unit_test() {
+        let cfg = Config::istanbul();
+
+        let compiled = COMPILED
+            .get("GreeterTest")
+            .expect("could not find contract");
+
+        let addr = "0x1000000000000000000000000000000000000000"
+            .parse()
+            .unwrap();
+        let state = initialize_contracts(vec![(addr, compiled.runtime_bytecode.clone())]);
+
+        let vicinity = new_vicinity();
+        let backend = new_backend(&vicinity, state);
+        let mut dapp = Executor::new(12_000_000, &cfg, &backend);
+
+        // call the setup function to deploy the contracts inside the test
+        let (_, status, _) = dapp
+            .call::<(), _>(
+                Address::zero(),
+                addr,
+                &get_func("function setUp() external").unwrap(),
+                (),
+                0.into(),
+            )
+            .unwrap();
+        assert_eq!(status, ExitReason::Succeed(ExitSucceed::Stopped));
+
+        let (_, status, _) = dapp
+            .call::<(), _>(
+                Address::zero(),
+                addr,
+                &get_func("function testGreeting()").unwrap(),
+                (),
+                0.into(),
+            )
+            .unwrap();
+        assert_eq!(status, ExitReason::Succeed(ExitSucceed::Stopped));
+    }
+
+    #[test]
+    fn failing_with_no_reason_if_no_setup() {
+        let cfg = Config::istanbul();
+
+        let compiled = COMPILED
+            .get("GreeterTest")
+            .expect("could not find contract");
+
+        let addr = "0x1000000000000000000000000000000000000000"
+            .parse()
+            .unwrap();
+        let state = initialize_contracts(vec![(addr, compiled.runtime_bytecode.clone())]);
+
+        let vicinity = new_vicinity();
+        let backend = new_backend(&vicinity, state);
+        let mut dapp = Executor::new(12_000_000, &cfg, &backend);
+
+        let (status, res) = dapp.executor.transact_call(
+            Address::zero(),
+            addr,
+            0.into(),
+            id("testFailGreeting()").to_vec(),
+            dapp.gas_limit,
+            vec![],
+        );
+        assert_eq!(status, ExitReason::Revert(ExitRevert::Reverted));
+        assert!(res.is_empty());
+    }
+
+    #[test]
+    fn failing_solidity_unit_test() {
+        let cfg = Config::istanbul();
+
+        let compiled = COMPILED
+            .get("GreeterTest")
+            .expect("could not find contract");
+
+        let addr = "0x1000000000000000000000000000000000000000"
+            .parse()
+            .unwrap();
+        let state = initialize_contracts(vec![(addr, compiled.runtime_bytecode.clone())]);
+
+        let vicinity = new_vicinity();
+        let backend = new_backend(&vicinity, state);
+        let mut dapp = Executor::new(12_000_000, &cfg, &backend);
+
+        // call the setup function to deploy the contracts inside the test
+        let (_, status, _) = dapp
+            .call::<(), _>(
+                Address::zero(),
+                addr,
+                &get_func("function setUp() external").unwrap(),
+                (),
+                0.into(),
+            )
+            .unwrap();
+        assert_eq!(status, ExitReason::Succeed(ExitSucceed::Stopped));
+
+        let (status, res) = dapp.executor.transact_call(
+            Address::zero(),
+            addr,
+            0.into(),
+            id("testFailGreeting()").to_vec(),
+            dapp.gas_limit,
+            vec![],
+        );
+        assert_eq!(status, ExitReason::Revert(ExitRevert::Reverted));
+        let reason = decode_revert(&res).unwrap();
+        assert_eq!(reason, "not equal to `hi`");
+    }
+}
