+++ dapp/src/multi_runner.rs
@@ -161,182 +161,7 @@ impl<'a> MultiContractRunner<'a> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{decode_revert, COMPILED};
-    use ethers::{
-        types::{H160, U256},
-        utils::id,
-    };
-
-    use dapp_utils::get_func;
-    use evm::{ExitReason, ExitRevert, ExitSucceed};
-
-    pub fn new_backend(vicinity: &MemoryVicinity, state: MemoryState) -> MemoryBackend<'_> {
-        MemoryBackend::new(vicinity, state)
-    }
-
-    pub fn new_vicinity() -> MemoryVicinity {
-        MemoryVicinity {
-            gas_price: U256::zero(),
-            origin: H160::default(),
-            block_hashes: Vec::new(),
-            block_number: Default::default(),
-            block_coinbase: Default::default(),
-            block_timestamp: Default::default(),
-            block_difficulty: Default::default(),
-            block_gas_limit: Default::default(),
-            chain_id: U256::one(),
-        }
-    }
-
-    #[test]
-    fn can_call_vm_directly() {
-        let cfg = Config::istanbul();
-        let compiled = COMPILED.get("Greeter").expect("could not find contract");
-
-        let addr = "0x1000000000000000000000000000000000000000"
-            .parse()
-            .unwrap();
-        let state = executor::initialize_contracts(vec![(addr, compiled.runtime_bytecode.clone())]);
-
-        let vicinity = new_vicinity();
-        let backend = new_backend(&vicinity, state);
-        let mut dapp = Executor::new(12_000_000, &cfg, &backend);
-
-        let (_, status, _) = dapp
-            .call::<(), _>(
-                Address::zero(),
-                addr,
-                &get_func("function greet(string greeting) external").unwrap(),
-                "hi".to_owned(),
-                0.into(),
-            )
-            .unwrap();
-        assert_eq!(status, ExitReason::Succeed(ExitSucceed::Stopped));
-
-        let (retdata, status, _) = dapp
-            .call::<String, _>(
-                Address::zero(),
-                addr,
-                &get_func("function greeting() public view returns (string)").unwrap(),
-                (),
-                0.into(),
-            )
-            .unwrap();
-        assert_eq!(status, ExitReason::Succeed(ExitSucceed::Returned));
-        assert_eq!(retdata, "hi");
-    }
-
-    #[test]
-    fn solidity_unit_test() {
-        let cfg = Config::istanbul();
-
-        let compiled = COMPILED
-            .get("GreeterTest")
-            .expect("could not find contract");
-
-        let addr = "0x1000000000000000000000000000000000000000"
-            .parse()
-            .unwrap();
-        let state = executor::initialize_contracts(vec![(addr, compiled.runtime_bytecode.clone())]);
-
-        let vicinity = new_vicinity();
-        let backend = new_backend(&vicinity, state);
-        let mut dapp = Executor::new(12_000_000, &cfg, &backend);
-
-        // call the setup function to deploy the contracts inside the test
-        let (_, status, _) = dapp
-            .call::<(), _>(
-                Address::zero(),
-                addr,
-                &get_func("function setUp() external").unwrap(),
-                (),
-                0.into(),
-            )
-            .unwrap();
-        assert_eq!(status, ExitReason::Succeed(ExitSucceed::Stopped));
-
-        let (_, status, _) = dapp
-            .call::<(), _>(
-                Address::zero(),
-                addr,
-                &get_func("function testGreeting()").unwrap(),
-                (),
-                0.into(),
-            )
-            .unwrap();
-        assert_eq!(status, ExitReason::Succeed(ExitSucceed::Stopped));
-    }
-
-    #[test]
-    fn failing_with_no_reason_if_no_setup() {
-        let cfg = Config::istanbul();
-
-        let compiled = COMPILED
-            .get("GreeterTest")
-            .expect("could not find contract");
-
-        let addr = "0x1000000000000000000000000000000000000000"
-            .parse()
-            .unwrap();
-        let state = executor::initialize_contracts(vec![(addr, compiled.runtime_bytecode.clone())]);
-
-        let vicinity = new_vicinity();
-        let backend = new_backend(&vicinity, state);
-        let mut dapp = Executor::new(12_000_000, &cfg, &backend);
-
-        let (status, res) = dapp.executor.transact_call(
-            Address::zero(),
-            addr,
-            0.into(),
-            id("testFailGreeting()").to_vec(),
-            dapp.gas_limit,
-            vec![],
-        );
-        assert_eq!(status, ExitReason::Revert(ExitRevert::Reverted));
-        assert!(res.is_empty());
-    }
-
-    #[test]
-    fn failing_solidity_unit_test() {
-        let cfg = Config::istanbul();
-
-        let compiled = COMPILED
-            .get("GreeterTest")
-            .expect("could not find contract");
-
-        let addr = "0x1000000000000000000000000000000000000000"
-            .parse()
-            .unwrap();
-        let state = executor::initialize_contracts(vec![(addr, compiled.runtime_bytecode.clone())]);
-
-        let vicinity = new_vicinity();
-        let backend = new_backend(&vicinity, state);
-        let mut dapp = Executor::new(12_000_000, &cfg, &backend);
-
-        // call the setup function to deploy the contracts inside the test
-        let (_, status, _) = dapp
-            .call::<(), _>(
-                Address::zero(),
-                addr,
-                &get_func("function setUp() external").unwrap(),
-                (),
-                0.into(),
-            )
-            .unwrap();
-        assert_eq!(status, ExitReason::Succeed(ExitSucceed::Stopped));
-
-        let (status, res) = dapp.executor.transact_call(
-            Address::zero(),
-            addr,
-            0.into(),
-            id("testFailGreeting()").to_vec(),
-            dapp.gas_limit,
-            vec![],
-        );
-        assert_eq!(status, ExitReason::Revert(ExitRevert::Reverted));
-        let reason = decode_revert(&res).unwrap();
-        assert_eq!(reason, "not equal to `hi`");
-    }
+    use crate::test_helpers::{new_backend, new_vicinity, COMPILED};
 
     #[test]
     fn test_runner() {
